+++ packages/business-rules-lib/src/validators/permission.js
@@ -1,12 +1,29 @@
+/**
+ * Validate a permission reference number.
+ *
+ * NOTE: This has been deliberately kept loose to allow for old-style licence numbers which used hex format for the last section
+ *
+ * @param joi
+ * @returns {this}
+ */
 export const createPermissionNumberValidator = joi =>
   joi
     .string()
     .trim()
-    .pattern(/^\d{8}-\d[A-Z]{2}\d[A-Z]{3}-[A-HJ-NP-Z0-9]{6}$/)
+    .uppercase()
+    .pattern(/^\d{8}-\d[A-Z]{2}\d[A-Z]{3}-[A-Z0-9]{6}$/)
     .required()
     .description('The permission reference number')
-    .example('17030621-3WC3FFT-U6HLG9')
+    .example('17030621-3WC3FFT-B6HLG9')
 
+/**
+ * Validate the last section of the permission reference number.
+ *
+ * NOTE: This has been deliberately kept loose to allow for old-style licence numbers which used hex format for the last section
+ *
+ * @param joi
+ * @returns {this}
+ */
 export const permissionNumberUniqueComponentValidator = joi =>
   joi
     .string()
@@ -14,5 +31,5 @@ export const permissionNumberUniqueComponentValidator = joi =>
     .uppercase()
     .pattern(/^[A-Z0-9]{6}$/)
     .required()
-    .description('The unqiue part of the permission reference number')
-    .example('U6HLG9')
+    .description('The unique part of the permission reference number')
+    .example('B6HLG9')